+++ src/TibiaWorldsWorldV3.go
@@ -4,6 +4,7 @@ import (
 	"log"
 	"regexp"
 	"strings"
+	"tibiadata-api-go/src/structs"
 
 	"github.com/PuerkitoBio/goquery"
 	"github.com/gin-gonic/gin"
@@ -11,47 +12,13 @@ import (
 
 // TibiaWorldsWorldV3 func
 func TibiaWorldsWorldV3(c *gin.Context) {
-
 	// getting params from URL
 	world := c.Param("world")
 
-	// Child of World
-	type OnlinePlayers struct {
-		Name     string `json:"name"`
-		Level    int    `json:"level"`
-		Vocation string `json:"vocation"`
-	}
-
-	// Child of Worlds
-	type World struct {
-		Name                string          `json:"name"`
-		Status              string          `json:"status"`                // Status:
-		PlayersOnline       int             `json:"players_online"`        // Players Online:
-		RecordPlayers       int             `json:"record_players"`        // Online Record:
-		RecordDate          string          `json:"record_date"`           // Online Record:
-		CreationDate        string          `json:"creation_date"`         // Creation Date: -> convert to YYYY-MM
-		Location            string          `json:"location"`              // Location:
-		PvpType             string          `json:"pvp_type"`              // PvP Type:
-		PremiumOnly         bool            `json:"premium_only"`          // Premium Type: premium = true / else: false
-		TransferType        string          `json:"transfer_type"`         // Transfer Type: regular (if not present) / locked / blocked
-		WorldsQuestTitles   []string        `json:"world_quest_titles"`    // World Quest Titles:
-		BattleyeProtected   bool            `json:"battleye_protected"`    // BattlEye Status: true if protected / false if "Not protected by BattlEye."
-		BattleyeDate        string          `json:"battleye_date"`         // BattlEye Status: null if since release / else show date?
-		GameWorldType       string          `json:"game_world_type"`       // Game World Type: regular / experimental / tournament (if Tournament World Type exists)
-		TournamentWorldType string          `json:"tournament_world_type"` // Tournament World Type: "" (default?) / regular / restricted
-		OnlinePlayers       []OnlinePlayers `json:"online_players"`
-	}
-
-	// Child of JSONData
-	type Worlds struct {
-		World World `json:"world"`
-	}
-
-	//
 	// The base includes two levels: World and Information
 	type JSONData struct {
-		Worlds      Worlds      `json:"worlds"`
-		Information Information `json:"information"`
+		World       structs.World       `json:"worlds"`
+		Information structs.Information `json:"information"`
 	}
 
 	// Adding fix for First letter to be upper and rest lower
@@ -68,15 +35,16 @@ func TibiaWorldsWorldV3(c *gin.Context) {
 	}
 
 	// Creating empty vars
-	var WorldsStatus, WorldsRecordDate, WorldsCreationDate, WorldsLocation, WorldsPvpType, WorldsTransferType, WorldsBattleyeDate, WorldsGameWorldType, WorldsTournamentWorldType string
-	var WorldsQuestTitles []string
-	var WorldsPlayersOnline, WorldsRecordPlayers int
-	var WorldsPremiumOnly, WorldsBattleyeProtected bool
-	var WorldsOnlinePlayers []OnlinePlayers
+	var (
+		WorldsStatus, WorldsRecordDate, WorldsCreationDate, WorldsLocation, WorldsPvpType, WorldsTransferType, WorldsBattleyeDate, WorldsGameWorldType, WorldsTournamentWorldType string
+		WorldsQuestTitles                                                                                                                                                         []string
+		WorldsPlayersOnline, WorldsRecordPlayers                                                                                                                                  int
+		WorldsPremiumOnly, WorldsBattleyeProtected                                                                                                                                bool
+		WorldsOnlinePlayers                                                                                                                                                       []structs.OnlinePlayer
+	)
 
 	// Running query over each div
 	ReaderHTML.Find(".Table1 .InnerTableContainer table tr").Each(func(index int, s *goquery.Selection) {
-
 		// Storing HTML into CreatureDivHTML
 		WorldsInformationDivHTML, err := s.Html()
 		if err != nil {
@@ -88,12 +56,12 @@ func TibiaWorldsWorldV3(c *gin.Context) {
 		subma1 := regex1.FindAllStringSubmatch(WorldsInformationDivHTML, -1)
 
 		if len(subma1) > 0 {
-
 			// Creating easy to use vars (and unescape hmtl right string)
 			WorldsInformationLeftColumn := subma1[0][1]
 			WorldsInformationRightColumn := TibiaDataSanitizeEscapedString(subma1[0][2])
 
-			if WorldsInformationLeftColumn == "Status" {
+			switch WorldsInformationLeftColumn {
+			case "Status":
 				switch {
 				case strings.Contains(WorldsInformationRightColumn, "</div>Online"):
 					WorldsStatus = "online"
@@ -102,11 +70,9 @@ func TibiaWorldsWorldV3(c *gin.Context) {
 				default:
 					WorldsStatus = "unknown"
 				}
-			}
-			if WorldsInformationLeftColumn == "Players Online" {
+			case "Players Online":
 				WorldsPlayersOnline = TibiadataStringToIntegerV3(WorldsInformationRightColumn)
-			}
-			if WorldsInformationLeftColumn == "Online Record" {
+			case "Online Record":
 				// Regex to get data for record values
 				regex2 := regexp.MustCompile(`(.*) players \(on (.*)\)`)
 				subma2 := regex2.FindAllStringSubmatch(WorldsInformationRightColumn, -1)
@@ -116,23 +82,17 @@ func TibiaWorldsWorldV3(c *gin.Context) {
 					WorldsRecordPlayers = TibiadataStringToIntegerV3(subma2[0][1])
 					WorldsRecordDate = TibiadataDatetimeV3(subma2[0][2])
 				}
-			}
-			if WorldsInformationLeftColumn == "Creation Date" {
+			case "Creation Date":
 				WorldsCreationDate = WorldsInformationRightColumn
-			}
-			if WorldsInformationLeftColumn == "Location" {
+			case "Location":
 				WorldsLocation = WorldsInformationRightColumn
-			}
-			if WorldsInformationLeftColumn == "PvP Type" {
+			case "PvP Type":
 				WorldsPvpType = WorldsInformationRightColumn
-			}
-			if WorldsInformationLeftColumn == "Premium Type" {
+			case "Premium Type":
 				WorldsPremiumOnly = true
-			}
-			if WorldsInformationLeftColumn == "Transfer Type" {
+			case "Transfer Type":
 				WorldsTransferType = WorldsInformationRightColumn
-			}
-			if WorldsInformationLeftColumn == "World Quest Titles" {
+			case "World Quest Titles":
 				if WorldsInformationRightColumn != "This game world currently has no title." {
 					WorldsQuestTitlesTmp := strings.Split(WorldsInformationRightColumn, ", ")
 					for _, str := range WorldsQuestTitlesTmp {
@@ -141,9 +101,7 @@ func TibiaWorldsWorldV3(c *gin.Context) {
 						}
 					}
 				}
-			}
-			if WorldsInformationLeftColumn == "BattlEye Status" {
-
+			case "BattlEye Status":
 				if WorldsInformationRightColumn == "Not protected by BattlEye." {
 					WorldsBattleyeProtected = false
 				} else {
@@ -156,11 +114,9 @@ func TibiaWorldsWorldV3(c *gin.Context) {
 						WorldsBattleyeDate = subma21[0][1]
 					}
 				}
-			}
-			if WorldsInformationLeftColumn == "Game World Type" {
+			case "Game World Type":
 				WorldsGameWorldType = strings.ToLower(WorldsInformationRightColumn)
-			}
-			if WorldsInformationLeftColumn == "Tournament World Type" {
+			case "Tournament World Type":
 				WorldsGameWorldType = "tournament"
 				if WorldsInformationRightColumn == "Restricted Store" {
 					WorldsTournamentWorldType = "restricted"
@@ -174,7 +130,6 @@ func TibiaWorldsWorldV3(c *gin.Context) {
 
 	// Running query over each div
 	ReaderHTML.Find(".Table2 .InnerTableContainer table tr").First().NextAll().Each(func(index int, s *goquery.Selection) {
-
 		// Storing HTML into CreatureDivHTML
 		WorldsInformationDivHTML, err := s.Html()
 		if err != nil {
@@ -186,8 +141,7 @@ func TibiaWorldsWorldV3(c *gin.Context) {
 		subma1 := regex1.FindAllStringSubmatch(WorldsInformationDivHTML, -1)
 
 		if len(subma1) > 0 {
-
-			WorldsOnlinePlayers = append(WorldsOnlinePlayers, OnlinePlayers{
+			WorldsOnlinePlayers = append(WorldsOnlinePlayers, structs.OnlinePlayer{
 				Name:     subma1[0][1],
 				Level:    TibiadataStringToIntegerV3(subma1[0][2]),
 				Vocation: subma1[0][3],
@@ -198,27 +152,25 @@ func TibiaWorldsWorldV3(c *gin.Context) {
 	//
 	// Build the data-blob
 	jsonData := JSONData{
-		Worlds: Worlds{
-			World{
-				Name:                world,
-				Status:              WorldsStatus,
-				PlayersOnline:       WorldsPlayersOnline,
-				RecordPlayers:       WorldsRecordPlayers,
-				RecordDate:          WorldsRecordDate,
-				CreationDate:        WorldsCreationDate,
-				Location:            WorldsLocation,
-				PvpType:             WorldsPvpType,
-				PremiumOnly:         WorldsPremiumOnly,
-				TransferType:        WorldsTransferType,
-				WorldsQuestTitles:   WorldsQuestTitles,
-				BattleyeProtected:   WorldsBattleyeProtected,
-				BattleyeDate:        WorldsBattleyeDate,
-				GameWorldType:       WorldsGameWorldType,
-				TournamentWorldType: WorldsTournamentWorldType,
-				OnlinePlayers:       WorldsOnlinePlayers,
-			},
+		World: structs.World{
+			Name:                world,
+			Status:              WorldsStatus,
+			PlayersOnline:       WorldsPlayersOnline,
+			RecordPlayers:       WorldsRecordPlayers,
+			RecordDate:          WorldsRecordDate,
+			CreationDate:        WorldsCreationDate,
+			Location:            WorldsLocation,
+			PvpType:             WorldsPvpType,
+			PremiumOnly:         WorldsPremiumOnly,
+			TransferType:        WorldsTransferType,
+			WorldsQuestTitles:   WorldsQuestTitles,
+			BattleyeProtected:   WorldsBattleyeProtected,
+			BattleyeDate:        WorldsBattleyeDate,
+			GameWorldType:       WorldsGameWorldType,
+			TournamentWorldType: WorldsTournamentWorldType,
+			OnlinePlayers:       WorldsOnlinePlayers,
 		},
-		Information: Information{
+		Information: structs.Information{
 			APIVersion: TibiadataAPIversion,
 			Timestamp:  TibiadataDatetimeV3(""),
 		},
