+++ src/structs/TibiaStructsV3.go
@@ -1,0 +1,7 @@
+package structs
+
+// Information stores some API data
+type Information struct {
+	APIVersion int    `json:"api_version"`
+	Timestamp  string `json:"timestamp"`
+}
